+++ compiler-tests/src/test/data/box/aggregation/MultipleBindingReplacementsFromSameClass.kt
@@ -1,3 +1,5 @@
+// GENERATE_CONTRIBUTION_PROVIDERS: false
+
 // MODULE: lib
 interface Foo